refactor(dev-dashboard): update onboarding intro
